+++ Chercheur/AjouterChercheurDialog.py
