+++ Laboratoire/AjouterLaboratoireDialog.py
