+++ Publication/BibliographieDialog.py
